// prisma/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const DocumentType = {
  PERSONAL_ID: "PERSONAL_ID",
  PASSPORT: "PASSPORT",
} as const

export type DocumentType = (typeof DocumentType)[keyof typeof DocumentType]